func-vbc-sfhc/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
cryptography==36.0.1
cygrpc
grpcio
azure-functions
boto3==1.21.27
botocore==1.24.27
FindSpark==2.0.0
Numpy
pandas
pyspark==3.2.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
salesforce-bulk==2.2.0
Unicodecsv==0.14.1
Urllib3==1.26.7
